LayoutTests/media/media-session/actionHandler-no-document-leak.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
<iframe id="iframe"></iframe>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/document-leak-test.js"></script>
<script>
description("Tests that page installed actionHandlers do not leak documents.");
onload = () => runDocumentLeakTest({ frameURL: "./resources/media-session-action-handler-document-leak-frame.html", framesToCreate: 20 });
</script>
</body>
</html>
